+++ client/src/components/Filtros/filtros.module.css
@@ -1,0 +1,5 @@
+.contenedorFiltros {
+    display: flex;
+    justify-content: space-evenly;
+    padding: 20px;
+  }
