+++ client/src/components/Paginado/Paginado.jsx
@@ -3,6 +3,7 @@ import { useEffect, useState } from "react";
 import estilo from './paginado.module.css'
 import Cards from '../../components/Cards/Cards'
 import { ObtenerPorPagina, paginaActual, userSearch, handleOrderCharacters, handleStatusCharacters, handleGenderCharacters } from '../../redux/actions';
+import Filtros from '../Filtros/filtros.jsx';
 const Paginado = () => {
 
     const dispatch = useDispatch();
@@ -12,7 +13,7 @@ const Paginado = () => {
     const userSearchG = useSelector(state => state.userSearchB);
     const [cantidadPorPag] = useState(20);
     const paginasDisponibles = []
-
+    
     for (let i = 0; i <= Math.ceil(Characters / cantidadPorPag); i++) {
         paginasDisponibles.push(i)
     }
@@ -41,47 +42,10 @@ const Paginado = () => {
         dispatch(ObtenerPorPagina(0));
         dispatch(userSearch(false));
     }
-
-    const handleOrder = (event) => {
-        event.preventDefault();
-        console.log("BANDERA DE MÉTODO HANDLEORDER:", event.target.value);
-        dispatch(handleOrderCharacters(event.target.value, currentPag));
-        dispatch(paginaActual(currentPag))
-    }
-
-    const handleStatus = (event) => {
-        event.preventDefault();
-        dispatch(handleStatusCharacters(event.target.value, currentPag));
-        dispatch(paginaActual(currentPag));
-    }
-
-    const handleGender = (event) => {
-        event.preventDefault();
-        dispatch(handleGenderCharacters(event.target.value, currentPag));
-        dispatch(paginaActual(currentPag));
-    }
-
     return (
         <>
-            <div className={estilo.contenedorFiltros}>
-                <select name="" id="" onChange={event => handleOrder(event)} >
-                    <option value="any">CUALQUIER ORDEN</option>
-                    <option value="asc">↑ ORDENAR DE A-Z</option>
-                    <option value="desc">↓ ORDENAR DE Z-A</option>
-                </select>
-                <select name="" id="" onChange={event => handleStatus(event)}>
-                    <option value="anyStatus">CUALQUIER ESTADO</option>
-                    <option value="alive">VIVO</option>
-                    <option value="dead">MUERTO</option>
-                    <option value="unknown">DESCONOCIDO</option>
-                </select>
-                <select name="" id="" onChange={event => handleGender(event)}>
-                    <option value="anyGender">CUALQUIER GÉNERO</option>
-                    <option value="Female">FEMENINO</option>
-                    <option value="Male">MASCULINO</option>
-                    <option value="Genderless">SIN GÉNERO</option>
-                    <option value="unknown">DESCONOCIDO</option>
-                </select>
+            <div>
+                <Filtros></Filtros>
             </div>
             <div className={estilo.contenedorCards}>
                 {
@@ -105,8 +69,6 @@ const Paginado = () => {
                         </>
                 }
             </div>
-
-
             <div>
                 <div className={estilo.contenedorCards}>
                     <Cards characters={personajesXPag} />
